Changes: -World-Realm-Player file naming scheme is now consistent thoughout the project. No more Terrain-World-Soul names. -Added some additional missing file checks for when savefiles get borked. -Began work on terrain V3 -Added new threads/counter class for counting among threads -Jobs now defaults to number of available cores on the system -Added new scene (gmsceneGame_6) for the game. Linked to scene name "game" -Fixed a bug in CHKAnimation that would prevent compilation on Win64 builds -Win32 build now compiled to use large page addresses -CMemoryMonitor now uses uint64_t for byte count -Debug realm and world _lucra/terra is now a valid World-Realm pair -Fixed issue in renderer where forward renderer would sometimes not render (noticeable with shadow cameras)

+++ oneEngine-core/source/core/system/CMemoryMonitorWin32.cpp
@@ -22,7 +22,7 @@ CMemoryMonitorWin32::CMemoryMonitorWin32 ( void )
 	process_handle = GetCurrentProcess();
 
 	iMaxStored = 1024*4;
-	p_setSize = new unsigned long [iMaxStored];
+	p_setSize = new uint64_t [iMaxStored];
 	for ( unsigned int i = 0; i < iMaxStored; i++ )
 	{
 		p_setSize[i] = 0;
@@ -51,17 +51,17 @@ void CMemoryMonitorWin32::Update ( void )
 }
 
 // == Getters ==
-unsigned long	CMemoryMonitorWin32::GetPeakUsage ( void )
+uint64_t	CMemoryMonitorWin32::GetPeakUsage ( void )
 {
-	unsigned long max = p_setSize[0];
-	for ( unsigned int i = 1; i < iMaxStored; i++ )
+	uint64_t max = p_setSize[0];
+	for ( uint64_t i = 1; i < iMaxStored; i++ )
 	{
 		if ( p_setSize[i] > max )
 			max = p_setSize[i];
 	}
 	return max;
 }
-unsigned long*	CMemoryMonitorWin32::GetUsage ( void )
+uint64_t*	CMemoryMonitorWin32::GetUsage ( void )
 {
 	return p_setSize;
 }
